pyproject.toml
[tool.poetry]
name = "find-my-bike"
version = "0.1.0"
description = ""
authors = ["Tilman Krokotsch <t.krokotsch@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.1"
webdriver-manager = "^3.5.2"
pytorch-lightning = "^1.5.9"
tqdm = "^4.62.3"
responses = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = {extras = ["d"], version = "^22.1.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
